@@ -2,14 +2,16 @@
 	"name": "icybee/module-editor",
 	"type": "icanboogie-module",
 	"description": "Provides an API to manage and use editors, as well as several editors.",
-	"keywords": [ "editor" ],
+	"keywords": [
+		"editor"
+	],
 	"license": "BSD-3-Clause",
 	"authors": [
 		{
 			"name": "Olivier Laviale",
 			"email": "olivier.laviale@gmail.com",
 			"homepage": "http://olvlvl.com/",
-			"role" : "Developer"
+			"role": "Developer"
 		}
 	],
 	"support": {
@@ -20,12 +22,22 @@
 	"prefer-stable": true,
 	"prefer-dist": true,
 	"require": {
-		"php": ">=5.5",
+		"php": ">=5.6",
 		"icanboogie/module-installer": "^1.2",
-		"icybee/core": "^3.0",
-		"icybee/module-images": "^3.0"
+		"icybee/core": "^4.0",
+		"icybee/module-images": "^4.0"
+	},
+	"require-dev": {
+		"icybee/patron": "^2.0"
 	},
 	"autoload": {
-		"classmap": [ "lib" ]
+		"classmap": [
+			"lib"
+		]
+	},
+	"autoload-dev": {
+		"classmap": [
+			"tests/Application.php"
+		]
 	}
 }
