@@ -4,6 +4,6 @@ namespace Icybee\Modules\Editor;
 
 return [
 
-	'ICanBoogie\Core::lazy_get_editors' => Collection::class . '::prototype_get_editors'
+	'ICanBoogie\Application::lazy_get_editors' => Collection::class . '::prototype_get_editors'
 
 ];
